+++ hibernate-core/src/main/java/org/hibernate/mapping/Fetchable.java
@@ -5,7 +5,6 @@
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
 package org.hibernate.mapping;
-import org.hibernate.FetchMode;
 
 /**
  * Any mapping with an outer-join attribute
