+++ hibernate-core/src/main/java/org/hibernate/mapping/ToOne.java
@@ -6,13 +6,11 @@
  */
 package org.hibernate.mapping;
 
-import org.hibernate.FetchMode;
 import org.hibernate.MappingException;
 import org.hibernate.boot.registry.classloading.spi.ClassLoaderService;
 import org.hibernate.boot.spi.MetadataBuildingContext;
 import org.hibernate.engine.spi.Mapping;
 import org.hibernate.internal.util.ReflectHelper;
-import org.hibernate.type.ComponentType;
 import org.hibernate.type.Type;
 
 import java.util.Objects;
