+++ hibernate-core/src/main/java/org/hibernate/persister/collection/AbstractCollectionPersister.java
@@ -20,7 +20,7 @@ import java.util.Set;
 import java.util.function.Consumer;
 
 import org.hibernate.AssertionFailure;
-import org.hibernate.FetchMode;
+import org.hibernate.mapping.FetchMode;
 import org.hibernate.Filter;
 import org.hibernate.HibernateException;
 import org.hibernate.LockOptions;
