+++ hibernate-core/src/main/java/org/hibernate/persister/collection/QueryableCollection.java
@@ -6,7 +6,7 @@
  */
 package org.hibernate.persister.collection;
 
-import org.hibernate.FetchMode;
+import org.hibernate.mapping.FetchMode;
 import org.hibernate.persister.entity.EntityPersister;
 import org.hibernate.persister.entity.Joinable;
 import org.hibernate.persister.entity.PropertyMapping;
