+++ hibernate-core/src/main/java/org/hibernate/persister/entity/OuterJoinLoadable.java
@@ -5,7 +5,7 @@
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
 package org.hibernate.persister.entity;
-import org.hibernate.FetchMode;
+import org.hibernate.mapping.FetchMode;
 import org.hibernate.engine.spi.CascadeStyle;
 import org.hibernate.type.EntityType;
 import org.hibernate.type.Type;
