+++ hibernate-core/src/test/java/org/hibernate/orm/test/fetchstrategyhelper/BatchFetchStrategyHelperTest.java
@@ -37,14 +37,14 @@ public class BatchFetchStrategyHelperTest extends BaseCoreFunctionalTestCase {
 	@Test
 	public void testManyToOneDefaultFetch() {
 		final AssociationType associationType = determineAssociationType( AnEntity.class, "otherEntityDefault" );
-		final org.hibernate.FetchMode fetchMode = determineFetchMode( AnEntity.class, "otherEntityDefault" );
-		assertSame( org.hibernate.FetchMode.JOIN, fetchMode );
+		final org.hibernate.mapping.FetchMode fetchMode = determineFetchMode( AnEntity.class, "otherEntityDefault" );
+		assertSame( org.hibernate.mapping.FetchMode.JOIN, fetchMode );
 		final FetchStyle fetchStyle = FetchOptionsHelper.determineFetchStyleByMetadata(
 				fetchMode,
 				associationType,
 				sessionFactory()
 		);
-		// batch size is ignored with org.hibernate.FetchMode.JOIN
+		// batch size is ignored with org.hibernate.mapping.FetchMode.JOIN
 		assertSame( FetchStyle.JOIN, fetchStyle );
 		final FetchTiming fetchTiming = FetchOptionsHelper.determineFetchTiming(
 				fetchStyle,
@@ -57,14 +57,14 @@ public class BatchFetchStrategyHelperTest extends BaseCoreFunctionalTestCase {
 	@Test
 	public void testManyToOneJoinFetch() {
 		final AssociationType associationType = determineAssociationType( AnEntity.class, "otherEntityJoin" );
-		final org.hibernate.FetchMode fetchMode = determineFetchMode( AnEntity.class, "otherEntityJoin" );
-		assertSame( org.hibernate.FetchMode.JOIN, fetchMode );
+		final org.hibernate.mapping.FetchMode fetchMode = determineFetchMode( AnEntity.class, "otherEntityJoin" );
+		assertSame( org.hibernate.mapping.FetchMode.JOIN, fetchMode );
 		final FetchStyle fetchStyle = FetchOptionsHelper.determineFetchStyleByMetadata(
 				fetchMode,
 				associationType,
 				sessionFactory()
 		);
-		// batch size is ignored with org.hibernate.FetchMode.JOIN
+		// batch size is ignored with org.hibernate.mapping.FetchMode.JOIN
 		assertSame( FetchStyle.JOIN, fetchStyle );
 		final FetchTiming fetchTiming = FetchOptionsHelper.determineFetchTiming(
 				fetchStyle,
@@ -77,8 +77,8 @@ public class BatchFetchStrategyHelperTest extends BaseCoreFunctionalTestCase {
 	@Test
 	public void testManyToOneSelectFetch() {
 		final AssociationType associationType = determineAssociationType( AnEntity.class, "otherEntitySelect" );
-		final org.hibernate.FetchMode fetchMode = determineFetchMode( AnEntity.class, "otherEntitySelect" );
-		assertSame( org.hibernate.FetchMode.SELECT, fetchMode );
+		final org.hibernate.mapping.FetchMode fetchMode = determineFetchMode( AnEntity.class, "otherEntitySelect" );
+		assertSame( org.hibernate.mapping.FetchMode.SELECT, fetchMode );
 		final FetchStyle fetchStyle = FetchOptionsHelper.determineFetchStyleByMetadata(
 				fetchMode,
 				associationType,
@@ -96,8 +96,8 @@ public class BatchFetchStrategyHelperTest extends BaseCoreFunctionalTestCase {
 	@Test
 	public void testCollectionDefaultFetch() {
 		final AssociationType associationType = determineAssociationType( AnEntity.class, "colorsDefault" );
-		final org.hibernate.FetchMode fetchMode = determineFetchMode( AnEntity.class, "colorsDefault" );
-		assertSame( org.hibernate.FetchMode.SELECT, fetchMode );
+		final org.hibernate.mapping.FetchMode fetchMode = determineFetchMode( AnEntity.class, "colorsDefault" );
+		assertSame( org.hibernate.mapping.FetchMode.SELECT, fetchMode );
 		final FetchStyle fetchStyle = FetchOptionsHelper.determineFetchStyleByMetadata(
 				fetchMode,
 				associationType,
@@ -115,14 +115,14 @@ public class BatchFetchStrategyHelperTest extends BaseCoreFunctionalTestCase {
 	@Test
 	public void testCollectionJoinFetch() {
 		final AssociationType associationType = determineAssociationType( AnEntity.class, "colorsJoin" );
-		final org.hibernate.FetchMode fetchMode = determineFetchMode( AnEntity.class, "colorsJoin" );
-		assertSame( org.hibernate.FetchMode.JOIN, fetchMode );
+		final org.hibernate.mapping.FetchMode fetchMode = determineFetchMode( AnEntity.class, "colorsJoin" );
+		assertSame( org.hibernate.mapping.FetchMode.JOIN, fetchMode );
 		final FetchStyle fetchStyle = FetchOptionsHelper.determineFetchStyleByMetadata(
 				fetchMode,
 				associationType,
 				sessionFactory()
 		);
-		// batch size is ignored with org.hibernate.FetchMode.JOIN
+		// batch size is ignored with org.hibernate.mapping.FetchMode.JOIN
 		assertSame( FetchStyle.JOIN, fetchStyle );
 		final FetchTiming fetchTiming = FetchOptionsHelper.determineFetchTiming(
 				fetchStyle,
@@ -135,8 +135,8 @@ public class BatchFetchStrategyHelperTest extends BaseCoreFunctionalTestCase {
 	@Test
 	public void testCollectionSelectFetch() {
 		final AssociationType associationType = determineAssociationType( AnEntity.class, "colorsSelect" );
-		final org.hibernate.FetchMode fetchMode = determineFetchMode( AnEntity.class, "colorsSelect" );
-		assertSame( org.hibernate.FetchMode.SELECT, fetchMode );
+		final org.hibernate.mapping.FetchMode fetchMode = determineFetchMode( AnEntity.class, "colorsSelect" );
+		assertSame( org.hibernate.mapping.FetchMode.SELECT, fetchMode );
 		final FetchStyle fetchStyle = FetchOptionsHelper.determineFetchStyleByMetadata(
 				fetchMode,
 				associationType,
@@ -154,8 +154,8 @@ public class BatchFetchStrategyHelperTest extends BaseCoreFunctionalTestCase {
 	@Test
 	public void testCollectionSubselectFetch() {
 		final AssociationType associationType = determineAssociationType( AnEntity.class, "colorsSubselect" );
-		final org.hibernate.FetchMode fetchMode = determineFetchMode( AnEntity.class, "colorsSubselect" );
-		assertSame( org.hibernate.FetchMode.SELECT, fetchMode );
+		final org.hibernate.mapping.FetchMode fetchMode = determineFetchMode( AnEntity.class, "colorsSubselect" );
+		assertSame( org.hibernate.mapping.FetchMode.SELECT, fetchMode );
 		final FetchStyle fetchStyle = FetchOptionsHelper.determineFetchStyleByMetadata(
 				fetchMode,
 				associationType,
@@ -171,7 +171,7 @@ public class BatchFetchStrategyHelperTest extends BaseCoreFunctionalTestCase {
 		assertSame( FetchTiming.DELAYED, fetchTiming );
 	}
 
-	private org.hibernate.FetchMode determineFetchMode(Class<?> entityClass, String path) {
+	private org.hibernate.mapping.FetchMode determineFetchMode(Class<?> entityClass, String path) {
 		OuterJoinLoadable entityPersister = (OuterJoinLoadable) sessionFactory().getEntityPersister( entityClass.getName() );
 		int index = ( (UniqueKeyLoadable) entityPersister ).getPropertyIndex( path );
 		return  entityPersister.getFetchMode( index );
