+++ hibernate-core/src/test/java/org/hibernate/orm/test/fetchstrategyhelper/NoProxyFetchStrategyHelperTest.java
@@ -34,8 +34,8 @@ public class NoProxyFetchStrategyHelperTest extends BaseCoreFunctionalTestCase {
 	@Test
 	public void testManyToOneDefaultFetch() {
 		final AssociationType associationType = determineAssociationType( AnEntity.class, "otherEntityDefault" );
-		final org.hibernate.FetchMode fetchMode = determineFetchMode( AnEntity.class, "otherEntityDefault" );
-		assertSame( org.hibernate.FetchMode.JOIN, fetchMode );
+		final org.hibernate.mapping.FetchMode fetchMode = determineFetchMode( AnEntity.class, "otherEntityDefault" );
+		assertSame( org.hibernate.mapping.FetchMode.JOIN, fetchMode );
 		final FetchStyle fetchStyle = FetchOptionsHelper.determineFetchStyleByMetadata(
 				fetchMode,
 				associationType,
@@ -53,8 +53,8 @@ public class NoProxyFetchStrategyHelperTest extends BaseCoreFunctionalTestCase {
 	@Test
 	public void testManyToOneJoinFetch() {
 		final AssociationType associationType = determineAssociationType( AnEntity.class, "otherEntityJoin" );
-		final org.hibernate.FetchMode fetchMode = determineFetchMode( AnEntity.class, "otherEntityJoin" );
-		assertSame( org.hibernate.FetchMode.JOIN, fetchMode );
+		final org.hibernate.mapping.FetchMode fetchMode = determineFetchMode( AnEntity.class, "otherEntityJoin" );
+		assertSame( org.hibernate.mapping.FetchMode.JOIN, fetchMode );
 		final FetchStyle fetchStyle = FetchOptionsHelper.determineFetchStyleByMetadata(
 				fetchMode,
 				associationType,
@@ -72,8 +72,8 @@ public class NoProxyFetchStrategyHelperTest extends BaseCoreFunctionalTestCase {
 	@Test
 	public void testManyToOneSelectFetch() {
 		final AssociationType associationType = determineAssociationType( AnEntity.class, "otherEntitySelect" );
-		final org.hibernate.FetchMode fetchMode = determineFetchMode( AnEntity.class, "otherEntitySelect" );
-		assertSame( org.hibernate.FetchMode.SELECT, fetchMode );
+		final org.hibernate.mapping.FetchMode fetchMode = determineFetchMode( AnEntity.class, "otherEntitySelect" );
+		assertSame( org.hibernate.mapping.FetchMode.SELECT, fetchMode );
 		final FetchStyle fetchStyle = FetchOptionsHelper.determineFetchStyleByMetadata(
 				fetchMode,
 				associationType,
@@ -89,7 +89,7 @@ public class NoProxyFetchStrategyHelperTest extends BaseCoreFunctionalTestCase {
 		assertSame( FetchTiming.IMMEDIATE, fetchTiming );
 	}
 
-	private org.hibernate.FetchMode determineFetchMode(Class<?> entityClass, String path) {
+	private org.hibernate.mapping.FetchMode determineFetchMode(Class<?> entityClass, String path) {
 		OuterJoinLoadable entityPersister = (OuterJoinLoadable) sessionFactory().getEntityPersister( entityClass.getName() );
 		int index = ( (UniqueKeyLoadable) entityPersister ).getPropertyIndex( path );
 		return  entityPersister.getFetchMode( index );
